server/requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.10.45
botocore==1.13.45
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
DateTime==4.3
docutils==0.15.2
Flask==1.1.1
gensim==3.8.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
multidict==4.7.2
nltk==3.4.5
numpy==1.18.0
pandas==0.25.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
scipy==1.4.1
six==1.13.0
smart-open==1.9.0
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.4.2
zope.interface==4.7.1
